Added Chat invitation dialog Changed avatar file names to hashes Profile dialog layout fixes (patch by Drugwash)
git-svn-id: http://miranda.googlecode.com/svn/trunk@8662 83df23c2-51fd-11de-bb24-2d5e9b4af41d

+++ miranda/protocols/AimOscar/tlv.cpp
@@ -71,7 +71,7 @@ char* TLV::part(int pos, int length)//returns part of the tlv value
 	value[length]='\0';
 	return value;
 }
-int TLV::whole(char* buf)//returns the whole tlv
+unsigned short TLV::whole(char* buf)//returns the whole tlv
 {
 	*(unsigned short*)buf = _htons(type_);
 	*(unsigned short*)&buf[2] = _htons(length_);
